Split query validation and MCP index option parsing
